+++ test/support/factory.ex
@@ -1,0 +1,4 @@
+defmodule Langtool.Factory do
+  use ExMachina.Ecto, repo: Langtool.Repo
+  use Langtool.UserFactory
+end